Moved out of the __init__.py the functions: flex_pokemon_collection()
